Remove hashmap allocation in load_dictionary
In discussion with Kevin, it turned out that it is to
be assumed that hashmap is already allocated outside
of the spell.c file. It is to be assumed the this can
then just be used as an array.

The greater difficulty with this design was the inability
to extract the allocated memory outside of load_dictionary
into the main file.